Update s3.md
The app generated file, `.forestry/settings.yml` exposes AWS credentials, and likely other 3rd party credentials. This is a security vulnerability that needs a fix.
One proposed solution is to instruct the user of Forestry.io to add the .forestry/settings.yml to their repo's `.gitignore` file.

I suspect that Forestry.io holds a table of credentials somewhere, and then writes those credentials to the repo. I think a better solution than instructing a user to update their .gitignore is to update the Forestry.io software to not commit keys/secrets to github.

https://www.datree.io/resources/secrets-management-aws#:~:text=2%20%E2%80%93%20NEVER%20commit%20secrets%20into%20your%20Git%20repositories,-When%20there%20are&text=Even%20worse%2C%20when%20the%20code,stolen%2C%20you%20will%20be%20compromised.

Please note that this vulnerability, if not addressed, can be a show stopper from using your platform but if it WERE addressed, makes you an extremely competitive offering because other markdown editing clients don't allow users to easily connect to S3.